Remove window-controls and use native ones
While transparent/overlaid title bars are extremely attractive, they
are impossible to get right in the current state of Tauri. The
window-controls package attempts to recreate the native design with
HTML/CSS, but obviously lacks native controls (snap layout, etc).

Tauri supports native overlaid title bars using titleBarStyle: Overlay
but this is not supported in Windows/Linux so using it is more painful
for maintenance than not.